Replace data types with fixed width typedefs (C99)
- Unified variable types (Closes #909)
short --> int16_t
unsigned short --> uint16_t
int --> int32_t
unsigned int --> uint32_t
long --> int32_t
unsigned long --> uint32_t
long long --> int64_t
unsigned long long --> uint64_t

- Added missing header includes

+++ src/stlink-lib/md5.c
@@ -5,6 +5,7 @@
  * This is free and unencumbered software released into the public domain - June 2013 - waterjuice.org
  */
 
+#include <stdint.h>
 #include <memory.h>
 
 #include "md5.h"
@@ -205,7 +206,7 @@ void Md5Update(Md5Context* Context /* [in out] */, void const* Buffer /* [in] */
     }
 
     if ( BufferSize >= 64 ) {
-        Buffer = TransformFunction( Context, Buffer, BufferSize & ~(unsigned long)0x3f );
+        Buffer = TransformFunction( Context, Buffer, BufferSize & ~(uint32_t)0x3f );
         BufferSize &= 0x3f;
     }
